mind/requirements.txt
pandas
matplotlib
jupyter
cvxpy
numpy
pytest
pytest-mock
nautilus_trader
openbb

torch
alpaca-py
